receivewindow tidied
